Split classes into respective files and use QThread for upload and download of file
